FE_Poly: move content from .templates.h to .cc
